Fix and improve mock config for openSUSE build
Fixes useradd command to not use -n that mock seems to use by default
but it does not exist in SUSE (or Fedora). Pass and use DIST_VER to
set the releasever and also simply use DIST as is in the root directory
name. The exact naming of the directory is not that important, as long
as it's unique for built dists.

Start using metalink and gpg key installed as part of mock for the SUSE
repo files. Also add qubes repos so that the mock build doesn't fail
to enable them when requested to be used in builder config. However,
any repo metadata fetch errors are ignored for now because qubes doesn't
have the repos in their server.

These changes bring the mock config in sync with both example openSUSE
config in mock as well as Fedora/CentOS mock config in qubes.